requirements.txt
streamlit>=1.42.0
authlib>=1.6.5
python-dotenv>=1.0.0